[PORT FROM IMIN_LEGACY] vold: fix 64 bit ioctl error
Changing the num_sectors used in ioctl with BLKGETSIZE because
the kernel expects an unsigned long type and then changes 64 bits
with a 64 bits userspace. This overwrites what's located close to
the parameter location if any.

Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-26507
Category: aosp improvement
Domain: Storage-Removable_Storage
Origin: internal
Upstream-Candidate: yes
Reviewed-on: https://android.intel.com:443/325430
Reviewed-on: https://android.intel.com:443/344239

+++ Fat.h
@@ -27,10 +27,10 @@ public:
                        int ownerUid, int ownerGid, int permMask,
                        bool createLost);
     static int check_extend(const char *fsPath, unsigned int numParts);
-    static int format(const char *fsPath, unsigned int numSectors, bool wipe);
+    static int format(const char *fsPath, unsigned long numSectors, bool wipe);
 
 private:
-    static void wipe(const char *fsPath, unsigned int numSectors);
+    static void wipe(const char *fsPath, unsigned long numSectors);
 };
 
 #endif
